pyproject.toml
[tool.poetry]
name = "US_STEEL_RFP"
authors = ["Samuel Mbugua <skm228@cornell.edu>"]
readme = "README.md"
version = "1.0"
description = "Data Extraction from Invoices"

[tool.poetry.dependencies]
python = ">=3.9"
azure-ai-ml = "1.18.0"
azure-core = "1.30.2"
azure-ai-formrecognizer = "3.3.3"
dynaconf = "3.2.4"
pandas = "2.2.2"
